sheltie: use /proc to find the bash binary
Because we have a shared pid namespace, /proc inside the container
matches /proc outside the container.  The special directory 'root'
inside a given process directory in /proc contains that process's view
of the filesystem.

We can thus locate the container's root filesystem by taking advantage
of the interpreter's /proc directory.

In order to make this work, we rely on the parent process existing in
the container's mount namespace, so that the path we construct maps
properly to the container's filesystem.

Signed-off-by: Samuel Karp <skarp@amazon.com>