Serial console support
This adds getty services for each serial console device on the host so
users can get to the admin container via serial console connections.

As part of this change, the sshd process is now managed by a
user-managed instance of systemd running in the admin container.

All the admin container user-managed services are "WantedBy" the
admin.target.

Users can specify a password hash for the console login in the admin
host container's user-data.