dockerfile: Remove /etc/config/selinux from image
The latest version of ec2-instance-connect has a dependency on the
selinux-policy and selinux-policy-targeted, which have a post-install
script that creates the /etc/config/selinux file within the container's
root filesystem to enforce SELinux. The libselinux library checks if
this file exists (among other things) to determine whether files
should be relabeled or not which is a problem with binaries in the
container's filesystem that have SELinux integration, since they will
attempt to relabel files using contexts provided by the container's
SELinux policies.

The SELinux contexts provided by the container's filesystem are not
valid since they are not present in the policy shipped in Bottlerocket.
Thus, any attempt to relabel the filesystems using contexts outside
the Bottlerocket's SELinux policy will result in an error. Deleting the
SELinux configuration file solves the problem, since the check described
above will fail.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>