use dedicated cgroup v1 hierarchy for systemd on cgroup v2 hosts
The admin container image is currently based on Amazon Linux 2 which
ships with systemd v219. This version of systemd only supports cgroup
v1. Since a user instance of systemd is run in the admin container, this
makes the admin container incompatible with a host that chose to use
cgroup v2 and will show in systemd refusing to start.

The problem can be worked around by making a cgroup v1 hierarchy
available to systemd in the admin container. This hierarchy will be
separate from the cgroup v2 hierarchy of the host, and only be mounted
inside the admin container. The user systemd process inside the
container is instructed to use that hierarchy via the downstream
`--cgroup-base` option. The start script takes care to not actually
mount any cgroup controllers, so the host-side systemd can manage
processes without interference.

An admin container running on a cgroup v1 host will not experience any
changes. An admin container running on a cgroup v2 host will have a
couple extra mounts under `/.bottlerocket/admin-container-cgroup` in its
mount namespace.

Signed-off-by: Markus Boehme <markubo@amazon.com>