systemd: ensure proper exit by resolving conflict between jobs
When exiting the admin container, the user instance of systemd logs the
following and skips proper clean-up:

    Requested transaction contains unmergeable jobs: Transaction contains conflicting jobs 'stop' and 'start' for systemd-exit.service. Probably contradicting requirement dependencies configured.
    Failed to enqueue exit.target job: Transaction contains conflicting jobs 'stop' and 'start' for systemd-exit.service. Probably contradicting requirement dependencies configured.

Exiting a user instance of systemd starts the `exit.target`, which depends
on `systemd-exit.service`, which in turn depends on `-.slice` (the root
slice) and `shutdown.target`. The stated conflict is between the latter
two, with `-.slice` defining a conflict with `shutdown.target`.

Resolve the conflict by backporting a change from upstream systemd that only
synthesizes the root slice. This ensures systemd in the container exits
properly, and e.g. removes empty cgroups.

Signed-off-by: Markus Boehme <markubo@amazon.com>